+++ src/xrGame/ai_space.cpp
@@ -101,7 +101,7 @@ void CAI_Space::load				(LPCSTR level_name)
 
 #ifdef DEBUG
 	Memory.mem_compact		();
-	u32						mem_usage = Memory.mem_usage();
+	size_t						mem_usage = Memory.mem_usage();
 	CTimer					timer;
 	timer.Start				();
 #endif
